+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Additional config file locations added. See README.md for details.
+- Support for basic auto completion in fish shell
 
 ## 1.1.0 - 2019-03-14